Moved checkout from component to page
