+++ pages/index.js
@@ -3,7 +3,6 @@ import DonutApp from "../components/DonutApp";
 import Menu from "../components/menu/Menu";
 import Order from "../components/order/Order";
 import AuthInfo from "../components/user/AuthInfo";
-import Checkout from "../components/order/Checkout";
 import OrderSummary from "../components/order/OrderSummary";
 import ZIPInfo from "../components/user/ZIPInfo";
